#151 C1: register Meta.permission_conditions without host _ir import
Hosts no longer import trusts.conditions._ir, so Core registers the
Meta option from trusts.apps via a private helper that does not import
trusts.core. Models that declare permission_conditions keep loading.
