+++ docs/test-coverage.md
@@ -73,7 +73,7 @@ These must keep passing in addition to the table above:
 - `LegacyContentRegistryDeletedTest` / `ConditionRegistryPreservedTest` / `HistoricalCompilerPreservedTest` / `LiveTerminalParityTest` / `DependentHostContributionTest` / `IsolatedAppsDoesNotDonateDependentContributionTest` / `DependentAuthorizationTest` / `UnknownTerminalFailsClosedTest` — issue #87 S7: static `Content._contents`, `is_content*`, `get_content_fieldlookup`, `filter_by_content`, `_get_trusts`, and the content-map half of `class_prepared` / `register_junction` are gone (runtime/behavioral structure, not source-token regex); `register_content` / `register_junction` walk Meta conditions only; Ticket `meta_own`, Trust `own`, Junction content conditions, duplicate/error handling, and system checks stay exact; `HistoricalGroupQueryCompiler` remains and mixin-only isolation is unchanged; Category/Ticket/Trust/Group allow/deny, direct/group-only, mixed-Trust, inactive/anonymous, `obj=None` unchanged; isolated host AppConfig `Ref` contributions for both documented dependent levels (ReceiptImage / ReceiptImageMeta equivalents) cover object `has_perm`, aggregate QS all-match, `common_permissions`, mixed-scope denial, and direct vs group-only; each grant/all-match/enumeration projection is one SQL independent of candidate count; unknown and undeclared manual dependents fail closed and never reach a deleted fallback. No schema or migration.
 - `AlongRegistrationTest` / `AlongIdentityRejectionTest` / `AlongShapeAndGraphTest` / `AlongToFieldAndUuidTest` / `AlongRuntimeGateTest` / `AlongE005Test` — issue #92: isolated noun-neutral models only. S/C/E metadata; terminal = walk-site; reverse-O2M, two-level reverse, reverse O2O, and forward multi-valued-inverse suffixes; non-PK `to_field`; integer/text/UUID identities; every rejected identity family plus mixed/mismatched/UUID-vs-text; frozen `register` first; zero mutation and zero SQL on registration failure; self-link, two-node cycle, uneven diamond, duplicate grants/paths, NULL/dangling, empty seed, bound cut; direct + recursive `OR` and multiple handles; shared-row; all three projections plus `all_match` at one SQL each; queryset filter before slicing; `EXPLAIN QUERY PLAN` materializes the generation walk once; actual-alias runtime gating; E005 with no aliases / supported / unsupported / failed probe / several aliases; isolated registry unscanned; non-recursive SQL and results unchanged. No schema or migration.
 - `ClosedPredicateRegistrationTest` / `ClosedPredicateAuthorizationTest` / `ClosedPredicateEqualToFieldAuthorizationTest` — issue #98: isolated GH-shaped models. Public `All` / `Equal` / `permission_in` plus the accepted team spelling (`user=t.team.members`, `condition=All(permission_in(...), Equal(...))`) register with **0 SQL**. Untyped / empty / mixed-root / incompatible / extra-multi-valued declarations fail closed without mutation. `Equal` paths that share a model but not one resolved comparison field (`to_field` slug vs code) fail closed with **0 SQL**; matching `to_field` registers and same-object allows / split objects deny at **1 SQL**. Team member + grant + bundle operation + same organization allows; removing membership, grant, bundle operation, or organization alignment denies only that branch; cross-org team grants deny. Direct and team roots OR; revoking one valid root preserves the other. Object, queryset/manager, and enumeration agree at **1 SQL**. Query construction is lazy; listings filter before pagination. Reverse O2M requester paths stay rejected. No schema or migration.
-- `ConstructTimeSelfBindTest` / `ImportOrderStandaloneConstructionTest` / `LiveEnsureSelfBindTest` / `OverrideUnbindAndFailClosedTest` / `NamedConditionCompilationTest` — issue #151 C1: `TrustsRegistry()` self-binds the private store adapter (`lookup.conditions is registry.conditions`); `freeze()` / `ready()` do not bind; live `_ensure` registries are bound without host `_ir` wiring; isolated subprocess import-order proofs (`trusts.core` then construct does not preload `_ir`; `_ir` then construct is not circular); explicit unbind and partial-bind non-mutation stay; unknown / unbound-record / malformed fail closed; construction / registration / `freeze()` are zero SQL; Core-only named-condition compilation through `record_for` / `compile_q` / `compile_registered_condition_q` / `evaluate_permission_condition`. Six-name `trusts.conditions` surface unchanged. No schema or migration.
+- `ConstructTimeSelfBindTest` / `ImportOrderStandaloneConstructionTest` / `LiveEnsureSelfBindTest` / `OverrideUnbindAndFailClosedTest` / `NamedConditionCompilationTest` — issue #151 C1: `TrustsRegistry()` self-binds the private store adapter (`lookup.conditions is registry.conditions`); `freeze()` / `ready()` do not bind; live `_ensure` registries are bound without host `_ir` wiring; isolated subprocess import-order proofs (`trusts.core` then construct does not preload `_ir`; `_ir` then construct is not circular; `trusts.apps` registers `Meta.permission_conditions` without loading `trusts.core` / `_ir`); explicit unbind and partial-bind non-mutation stay; unknown / unbound-record / malformed fail closed; construction / registration / `freeze()` are zero SQL; Core-only named-condition compilation through `record_for` / `compile_q` / `compile_registered_condition_q` / `evaluate_permission_condition`. Six-name `trusts.conditions` surface unchanged. No schema or migration.
 - `KernelConfigTest` / `LegacyMatrixATest` / `ConditionLookupBindTest` / `ConditionLookupBackendAdapterTest` / `AuthorizedQuerySetSurfaceTest` / `AuthorizedQuerySetLiveTest` / `FilterAuthorizedScopesIsolatedTest` / `FilterAuthorizedScopesToFieldTest` / `FilterAuthorizedScopesLiveTest` — historical issue #54 coverage (superseded by the later core/Zero split): `kernel_config()` then resolved `get_app_config('trusts')` by class identity; matrix A still populates `label='trusts'` with `trusts.0001_initial` / `0002_trustgroup` from `trusts.migrations` and no Zero package; `ConditionLookup` bind is zero-SQL, missing methods do not partial-bind, callables are never invoked; unbound overlay preserves `Content` conditions; `AuthorizedQuerySet.authorized` rejects strings with zero SQL, does not call `is_active_principal` / `get_permission`, AND-overlays `extra_q`, and has no `.permitted` / `.get_permission`; `filter_authorized_scopes` fail-closes unknown prefix / empty handles / content-terminal queryset / wrong permission type with zero SQL, correlates prefix hops to the resolved target field (non-PK `to_field` slug identity, sibling deny, lazy, one SQL), and does not use historical TrustGroup (live `filter_by_user_content_perm` still does). No schema, label, or migration change. `trust_grant_q` / `HistoricalGroupQueryCompiler` remain.
 
 Do not treat a gap as license to widen access. New grants need an explicit
